abstracted user jwt response for reusability
