+++ backend/src/helpers/userWithJwtResponse.ts
@@ -1,0 +1,24 @@
+import { User } from '../entities/User';
+import JwtService from '../services/JwtService';
+
+export type UserWithJWTResponse = {
+  jwt: string;
+  user: {
+    id: string;
+    twitterUserId: string;
+    screenName: string;
+  };
+};
+const JwtResponseFromUser = (user: User): UserWithJWTResponse => {
+  const jwt = JwtService.fromUser(user);
+  return {
+    jwt,
+    user: {
+      id: user.id,
+      twitterUserId: user.twitterUserId,
+      screenName: user.screenName,
+    },
+  };
+};
+
+export default JwtResponseFromUser;
